+++ packages/netlify-cms-locales/src/cs/index.js
@@ -125,6 +125,14 @@ const cs = {
       unknownPreview: {
         noPreview: "Žádný náhled pro widget '%{widget}'.",
       },
+      headingOptions: {
+        headingOne: 'Heading 1',
+        headingTwo: 'Heading 2',
+        headingThree: 'Heading 3',
+        headingFour: 'Heading 4',
+        headingFive: 'Heading 5',
+        headingSix: 'Heading 6',
+      },
     },
   },
   mediaLibrary: {
@@ -190,6 +198,7 @@ const cs = {
       newPost: 'Nový post',
       description:
         '%{smart_count} záznam čeká na schválení, %{readyCount} připraven k publikaci. |||| %{smart_count} čeká na schválení, %{readyCount} připraveno k publikaci. ',
+      dateFormat: 'MMMM D',
     },
     workflowCard: {
       lastChange: '%{date} (%{author})',