+++ packages/netlify-cms-proxy-server/src/middlewares/common/index.ts
@@ -1,0 +1,9 @@
+import express from 'express';
+import morgan from 'morgan';
+import cors from 'cors';
+
+export const registerCommonMiddlewares = (app: express.Express) => {
+  app.use(morgan('combined'));
+  app.use(cors());
+  app.use(express.json({ limit: '50mb' }));
+};