+++ packages/netlify-cms-ui-legacy/src/styles.js
@@ -321,6 +321,7 @@ const components = {
     display: flex;
     justify-content: space-between;
     align-items: center;
+    min-width: max-content;
 
     &:last-of-type {
       border-bottom: 0;