+++ packages/netlify-cms-widget-markdown/src/MarkdownControl/ToolbarButton.js
@@ -23,22 +23,16 @@ const StyledToolbarButton = styled.button`
   }
 `;
 
-const ToolbarButton = ({ type, label, icon, onClick, isActive, isHidden, disabled }) => {
-  if (isHidden) {
-    return null;
-  }
-
-  return (
-    <StyledToolbarButton
-      isActive={isActive}
-      onClick={e => onClick && onClick(e, type)}
-      title={label}
-      disabled={disabled}
-    >
-      {icon ? <Icon type={icon} /> : label}
-    </StyledToolbarButton>
-  );
-};
+const ToolbarButton = ({ type, label, icon, onClick, isActive, disabled }) => (
+  <StyledToolbarButton
+    isActive={isActive}
+    onClick={e => onClick && onClick(e, type)}
+    title={label}
+    disabled={disabled}
+  >
+    {icon ? <Icon type={icon} /> : label}
+  </StyledToolbarButton>
+);
 
 ToolbarButton.propTypes = {
   type: PropTypes.string,
@@ -46,7 +40,6 @@ ToolbarButton.propTypes = {
   icon: PropTypes.string,
   onClick: PropTypes.func,
   isActive: PropTypes.bool,
-  isHidden: PropTypes.bool,
   disabled: PropTypes.bool,
 };